Renamed tests to be in uniform location
