+++ tests/Adapter/IndexTestCase.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Phpactor\Indexer\Tests\Integration;
+namespace Phpactor\Indexer\Tests\Adapter;
 
 use Phpactor\Indexer\Model\Indexer;
 use Phpactor\Indexer\Tests\IntegrationTestCase;
