+++ tests/Adapter/Manifest/buildIndex.php.test
