+++ tests/Adapter/Tolerant/Indexer/ClassDeclarationIndexerTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Phpactor\Indexer\Tests\Integration\Tolerant\Indexer;
+namespace Phpactor\Indexer\Tests\Adapter\Tolerant\Indexer;
 
 use Generator;
 use Phpactor\Indexer\Adapter\Tolerant\Indexer\ClassDeclarationIndexer;
 use Phpactor\Indexer\Model\Record\ClassRecord;
-use Phpactor\Indexer\Tests\Integration\Tolerant\TolerantIndexerTestCase;
+use Phpactor\Indexer\Tests\Adapter\Tolerant\TolerantIndexerTestCase;
 use RuntimeException;
 
 class ClassDeclarationIndexerTest extends TolerantIndexerTestCase
