+++ tests/Adapter/Tolerant/TolerantIndexBuilderTest.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Phpactor\Indexer\Tests\Adapter\Tolerant;
+
+use Phpactor\Indexer\Tests\Adapter\IndexBuilderTestCase;
+
+class TolerantIndexBuilderTest extends IndexBuilderTestCase
+{
+}
