--- tests/Integration/Tolerant/TolerantIndexBuilderTest.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Phpactor\Indexer\Tests\Integration\Tolerant;
-
-use Phpactor\Indexer\Tests\Integration\IndexBuilderTestCase;
-
-class TolerantIndexBuilderTest extends IndexBuilderTestCase
-{
-}
